Added the reservation to the traveler reservations list + updated numseats of ride

+++ src/main/java/eus/ehu/ridesfx/domain/Ride.java
@@ -170,7 +170,7 @@ public class Ride implements Serializable {
 	 *
 	 * @return the available places
 	 */
-	public float getNumPlaces() {
+	public int getNumPlaces() {
 		return numPlaces;
 	}
